requirements.txt
pywebview>=4.0.0
pyinstaller>=5.0.0
Pillow>=9.0.0  # 可选：用于创建图标
python-dateutil>=2.8.0  # 用于日期计算